@@ -11,48 +11,71 @@
     "dev": "NODE_ENV=development nodemon src/app/index.ts",
     "start:dev": "NODE_ENV=development node .",
     "start:prod": "NODE_ENV=production node .",
+<<<<<<< Updated upstream
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
     "test": "jest",
     "test:coverage": "jest --coverage"
+=======
+    "lint": "eslint ./src/**/*.ts",
+    "lint:fix": "eslint ./src/**/*.ts --fix",
+    "format": "prettier --check ./src/**/*.ts",
+    "format:fix": "prettier --write ./src/**/*.ts",
+    "prepare": "husky",
+    "test": "jest"
+>>>>>>> Stashed changes
   },
   "dependencies": {
-    "@hapi/boom": "^9.1.4",
-    "bcrypt": "^5.0.1",
-    "cloudinary": "^1.28.1",
+    "@hapi/boom": "^10.0.1",
+    "bcrypt": "^5.1.1",
+    "cloudinary": "^2.4.0",
     "cors": "^2.8.5",
+<<<<<<< Updated upstream
     "dotenv": "^16.0.0",
     "express": "^4.17.3",
     "helmet": "^6.0.0",
+=======
+    "express": "^4.19.2",
+    "helmet": "^7.1.0",
+>>>>>>> Stashed changes
     "jsonwebtoken": "^9.0.2",
     "jwt-decode": "^3.1.2",
-    "mongodb": "^6.3.0",
-    "mongoose": "^8.1.0",
+    "mongodb": "^6.8.1",
+    "mongoose": "^8.6.1",
     "multer": "^1.4.4",
     "multer-storage-cloudinary": "^4.0.0",
-    "rollbar": "^2.26.1"
+    "rollbar": "^2.26.4"
   },
   "devDependencies": {
     "@jest/globals": "^29.7.0",
-    "@nightwatch/apitesting": "^3.1.3",
-    "@types/bcrypt": "^5.0.0",
-    "@types/cors": "^2.8.12",
-    "@types/express": "^4.17.13",
-    "@types/jest": "^29.5.11",
-    "@types/jsonwebtoken": "^8.5.8",
-    "@types/morgan": "^1.9.3",
-    "@types/multer": "^1.4.7",
-    "@types/node": "^17.0.8",
+    "@nightwatch/apitesting": "^3.1.4",
+    "@types/bcrypt": "^5.0.2",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/jest": "^29.5.12",
+    "@types/jsonwebtoken": "^8.5.9",
+    "@types/morgan": "^1.9.9",
+    "@types/multer": "^1.4.12",
+    "@types/node": "^17.0.45",
     "@types/supertest": "^6.0.2",
-    "@typescript-eslint/eslint-plugin": "^5.36.2",
-    "@typescript-eslint/parser": "^5.36.2",
-    "eslint": "^8.23.0",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
+    "@typescript-eslint/parser": "^5.62.0",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-security": "^3.0.1",
+    "eslint-plugin-sonarjs": "^2.0.2",
+    "husky": "^9.1.5",
     "jest": "^29.7.0",
+<<<<<<< Updated upstream
     "jest-html-reporters": "^3.1.7",
     "jest-junit": "^16.0.0",
     "nodemon": "^3.0.3",
+=======
+    "nodemon": "^3.1.4",
+    "prettier": "^3.3.3",
+>>>>>>> Stashed changes
     "supertest": "^6.3.4",
-    "ts-jest": "^29.1.1",
-    "ts-node": "^10.5.0",
-    "typescript": "^5.3.3"
+    "ts-jest": "^29.2.5",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.5.4"
   }
 }
