@@ -1,10 +1,10 @@
-import { app } from "./main";
-import { loadDatabase } from "./db";
-import Config from "./config/config";
+import { app } from './main';
+import { loadDatabase } from './db';
+import Config from './config/config';
 
 app.listen(Config.app.port || 3000, async () => {
-    await loadDatabase(Config.db.uri);
-    console.log(`
+  await loadDatabase(Config.db.uri);
+  console.log(`
         ██╗    ██╗ █████╗ ██╗███████╗██╗   ██╗██╗      █████╗ ███╗   ██╗██████╗      █████╗ ██████╗ ██╗
         ██║    ██║██╔══██╗██║██╔════╝██║   ██║██║     ██╔══██╗████╗  ██║██╔══██╗    ██╔══██╗██╔══██╗██║
         ██║ █╗ ██║███████║██║█████╗  ██║   ██║██║     ███████║██╔██╗ ██║██║  ██║    ███████║██████╔╝██║
