@@ -1,16 +1,16 @@
 export const hasTag = (id: number | undefined) => {
-  if (id) {
-    return {
-      path: "tag",
-      select: "-_id",
-      match: {
-        tag_id: id,
-      },
-    };
-  } else {
+  if (!id) {
     return {
       path: "tag",
       select: "-_id",
     };
   }
+
+  return {
+    path: "tag",
+    select: "-_id",
+    match: {
+      tag_id: id,
+    },
+  };
 };
