@@ -1,8 +1,8 @@
 // Internal Modules
-import Image from "./schema/image-schema";
-import Tag from "../tag/schema/tag-schema";
-import { hasTag } from "../common/utils/ref";
-import { IImage } from "./interfaces/image-interface";
+import Image from './schema/image-schema';
+import Tag from '../tag/schema/tag-schema';
+import { hasTag } from '../common/utils/ref';
+import { IImage } from './interfaces/image-interface';
 
 class ImageRepository {
   /**
@@ -10,7 +10,6 @@ class ImageRepository {
    * @param {IImage} image - the image to create in the database
    * @return { Promise<IImage> } - A new image created
    */
-
   async create(image: IImage): Promise<IImage> {
     const tagExists = await Tag.findOne({ tag_id: image.tag.tag_id });
     const _idTag = tagExists?._id;
@@ -25,10 +24,9 @@ class ImageRepository {
    * @param tag_id - the id from the tag to retrieve
    * @return { Promise<IImage[]> } An array of images
    */
-
   async findImages(tag_id?: number): Promise<IImage[]> {
     const images = await Image.find().populate(hasTag(tag_id));
-    return images.filter((image) => image.tag !== null);
+    return images.filter(image => image.tag !== null);
   }
 }
 
