@@ -1,12 +1,12 @@
 // External Modules
-import { Router } from "express";
+import { Router } from 'express';
 
 // Internal Modules
-import tagController from "./tag-controller";
+import tagController from './tag-controller';
 
 const tagRouter = Router();
 
-tagRouter.get("/", tagController.getTags);
-tagRouter.get("/:id", tagController.getTagsId);
+tagRouter.get('/', tagController.getTags);
+tagRouter.get('/:id', tagController.getTagsId);
 
 export default tagRouter;
